+++ src/EventSubscriber/DependencyUpdateSubscriber.php
@@ -2,15 +2,18 @@
 
 namespace Drupal\wmpathauto\EventSubscriber;
 
+use Drupal\Core\Config\ConfigCrudEvent;
+use Drupal\Core\Config\ConfigEvents;
 use Drupal\Core\KeyValueStore\KeyValueFactoryInterface;
 use Drupal\Core\KeyValueStore\KeyValueStoreInterface;
 use Drupal\Core\Queue\QueueFactory;
+use Drupal\hook_event_dispatcher\Event\Entity\EntityUpdateEvent;
 use Drupal\hook_event_dispatcher\Event\Path\PathUpdateEvent;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
 use Drupal\wmpathauto\Plugin\QueueWorker\AliasQueueWorker;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
-class PathUpdateSubscriber implements EventSubscriberInterface
+class DependencyUpdateSubscriber implements EventSubscriberInterface
 {
     /** @var KeyValueFactoryInterface */
     protected $keyValueFactory;
@@ -27,15 +30,39 @@ class PathUpdateSubscriber implements EventSubscriberInterface
 
     public static function getSubscribedEvents()
     {
-        $events[HookEventDispatcherInterface::PATH_UPDATE][] = ['updateDependentAliases'];
+        $events[HookEventDispatcherInterface::PATH_UPDATE][] = ['onPathUpdate'];
+        $events[ConfigEvents::SAVE][] = ['onConfigUpdate'];
+        $events[HookEventDispatcherInterface::ENTITY_UPDATE][] = ['onEntityUpdate'];
 
         return $events;
     }
 
-    public function updateDependentAliases(PathUpdateEvent $event): void
+    public function onPathUpdate(PathUpdateEvent $event): void
+    {
+        $this->queueDependents('pid:' . $event->getPid());
+    }
+
+    public function onConfigUpdate(ConfigCrudEvent $event): void
+    {
+        $this->queueDependents('config:' . $event->getConfig()->getName());
+    }
+
+    public function onEntityUpdate(EntityUpdateEvent $event): void
+    {
+        $entity = $event->getEntity();
+        $suffix = implode(':', [
+            'entity',
+            $entity->getEntityTypeId(),
+            $entity->id(),
+        ]);
+
+        $this->queueDependents($suffix);
+    }
+
+    protected function queueDependents(string $suffix): void
     {
         $queue = $this->queueFactory->get(AliasQueueWorker::ID);
-        $storage = $this->getStorage('pid:' . $event->getPid());
+        $storage = $this->getStorage($suffix);
 
         // Queue all entities depending on this pid
         foreach ($storage->getAll() as $dependent) {