+++ src/EventSubscriber/PatternDependenciesSubscriber.php
@@ -3,6 +3,7 @@
 namespace Drupal\wmpathauto\EventSubscriber;
 
 use Drupal\Core\Entity\ContentEntityInterface;
+use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\KeyValueStore\KeyValueFactoryInterface;
 use Drupal\Core\KeyValueStore\KeyValueStoreInterface;
 use Drupal\pathauto\PathautoPatternInterface;
@@ -47,29 +48,52 @@ class PatternDependenciesSubscriber implements EventSubscriberInterface
 
         $dependencies = $this->dependencyResolver->getDependencies($pattern, $entity);
 
-        foreach ($dependencies as $pid) {
-            $storage = $this->getStorage('pid:' . $pid);
-
-            $cid = sprintf(
-                '%s.%s.%s',
-                $entity->getEntityTypeId(),
-                $entity->id(),
-                $entity->language()->getId()
-            );
-
-            if ($storage->has($cid)) {
-                continue;
-            }
-
-            $storage->set(
-                $cid,
-                [
-                    'entityTypeId' => $entity->getEntityTypeId(),
-                    'entityId' => $entity->id(),
-                    'langcode' => $entity->language()->getId(),
-                ]
-            );
+        foreach ($dependencies->getAliases() as $pid) {
+            $suffix = implode(':', ['pid', $pid]);
+            $this->addEntityAsDependency($entity, $suffix);
         }
+
+        foreach ($dependencies->getConfigs() as $dependentConfig) {
+            $suffix = implode(':', [
+                'config',
+                $dependentConfig->getName(),
+            ]);
+            $this->addEntityAsDependency($entity, $suffix);
+        }
+
+        foreach ($dependencies->getEntities() as $dependentEntity) {
+            $suffix = implode(':', [
+                'entity',
+                $dependentEntity->getEntityTypeId(),
+                $dependentEntity->id(),
+            ]);
+            $this->addEntityAsDependency($entity, $suffix);
+        }
+    }
+
+    protected function addEntityAsDependency(EntityInterface $entity, string $suffix): void
+    {
+        $storage = $this->getStorage($suffix);
+
+        $cid = sprintf(
+            '%s.%s.%s',
+            $entity->getEntityTypeId(),
+            $entity->id(),
+            $entity->language()->getId()
+        );
+
+        if ($storage->has($cid)) {
+            return;
+        }
+
+        $storage->set(
+            $cid,
+            [
+                'entityTypeId' => $entity->getEntityTypeId(),
+                'entityId' => $entity->id(),
+                'langcode' => $entity->language()->getId(),
+            ]
+        );
     }
 
     protected function getStorage(string $suffix): KeyValueStoreInterface