+++ src/PatternDependencyCollectionInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Drupal\wmpathauto;
+
+use Drupal\Core\Config\Config;
+use Drupal\Core\Entity\EntityInterface;
+
+interface PatternDependencyCollectionInterface
+{
+    public function getAliases(): array;
+
+    public function getEntities(): array;
+
+    public function getConfigs(): array;
+
+    public function addPathAlias(string $pid): void;
+
+    public function addEntity(EntityInterface $entity): void;
+
+    public function addConfig(Config $config): void;
+}