+++ src/PatternDependencyResolverInterface.php
@@ -7,11 +7,6 @@ use Drupal\pathauto\PathautoPatternInterface;
 
 interface PatternDependencyResolverInterface
 {
-    /**
-     * Collect dependencies of a pathauto pattern generated for a certain entity.
-     *
-     * @return array
-     *      An array of pids (path alias identifiers)
-     */
-    public function getDependencies(PathautoPatternInterface $pattern, EntityInterface $entity): array;
+    /** Collect dependencies of a pathauto pattern generated for a certain entity. */
+    public function getDependencies(PathautoPatternInterface $pattern, EntityInterface $entity): PatternDependencyCollectionInterface;
 }